pyproject.toml
[project]
name = "polymarket-agent"
version = "0.1.0"
description = "Automated football arbitrage scanner for Polymarket"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "pydantic>=2.5",
    "pydantic-settings>=2.1",
    "pyyaml>=6.0",
    "apscheduler>=3.10,<4.0",
    "structlog>=24.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "respx>=0.21",
    "ruff>=0.3",
]

[project.scripts]
polymarket-scan = "src.main:main"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
